Rewrite into code_climate.Client
This is a more structured approach helping make it easier to perform
the increased required API calls for paging the issues.

A note on builds:
Code climate seems to cache a build for a commit sha, so they will
reuse an existing snapshot for a commit sha instead of stareting a new
build. So while it is important to make sure that the the latest build
is completed, it is not a good measure of what contains the code climate
measurements for the latest push to the main branch.